rearranged endpoints and methods between admin and problem controller made it more restful
