+++ Services/MainApp/MainApp.Api/Controllers/AdminController.cs
@@ -5,6 +5,7 @@ using Microsoft.AspNetCore.Mvc;
 
 namespace MainApp.Api.Controllers;
 
+[Route("api/admin")]
 [ApiController]
 
 [Authorize(AuthenticationSchemes = "Bearer")]
@@ -18,64 +19,10 @@ public class AdminController : ControllerBase
         _adminService = adminService;
     }
 
-    [HttpGet("/get-users")]
+    [HttpGet("users")]
     public async Task<IActionResult> GetUsers()
     {
        return Ok(await _adminService.GetUsers());
     }
     
-   [HttpGet("/get-problem/{id}")]
-    public async Task<IActionResult> GetProblem(int id)
-    {
-        var problem = await _adminService.GetProblem(id);
-        
-        return Ok(problem);
-    }
-    
-    [HttpPost("/add-problem")]
-    public async Task<IActionResult> AddProblem(AddProblemDto problem)
-    { 
-        try
-        {
-            await _adminService.AddProblem(problem);
-            
-            return Created();
-        }
-        catch (Exception ex)
-        {
-            return StatusCode(StatusCodes.Status500InternalServerError, ex.Message);
-        }
-        
-    }
-
-
-    [HttpPut("/edit-problem/{id}")]
-    public async Task<IActionResult> EditProblem(int id, AddProblemDto editProblem)
-    {
-        try
-        {
-            await _adminService.EditProblem(id, editProblem);
-        }
-        catch (Exception ex)
-        {
-            return StatusCode(StatusCodes.Status500InternalServerError, ex.Message);
-        }
-        
-        return Accepted();
-    }
-
-    [HttpDelete("/delete-problem/{id}")]
-    public async Task<IActionResult> DeleteProblem(int id)
-    {
-        try
-        {
-            await _adminService.DeleteProblem(id);
-        }
-        catch (Exception ex)
-        {
-            return StatusCode(StatusCodes.Status500InternalServerError, ex.Message);
-        }
-
-        return Ok();
-    }
 }
