+++ Services/MainApp/MainApp.Application/Services/ProblemsService.cs
@@ -1,6 +1,7 @@
 ﻿using BuildingBlocks.Common.Dtos;
 using BuildingBlocks.Common.Enums;
 using BuildingBlocks.Common.Helpers;
+using MainApp.Application.Dto.Request;
 using MainApp.Domain.Entity;
 using MainApp.Infrastructure.Data;
 using Microsoft.EntityFrameworkCore;
@@ -77,6 +78,102 @@ public class ProblemsService
         
         return problemResponse;
     }
+    
+         public async Task AddProblem(AddProblemDto problemDto)
+     {
+
+         var parsedCategory = ParseCategories(problemDto.Categories);
+         
+         var parsedDifficulty = ParseDifficulty(problemDto.Difficulty);
+         
+        
+        var problem = new Problem
+        {
+        
+            Name = problemDto.Name,
+            ProblemText = problemDto.ProblemText,
+            Category =  parsedCategory,
+            Difficulty = parsedDifficulty,
+            RuntimeLimit = problemDto.RuntimeLimit,
+            MemoryLimit = problemDto.MemoryLimit,
+            TestCases = problemDto.TestCases.Select(tc => new TestCase
+            {
+                Input = tc.Input,
+                ExpectedOutput = tc.ExpectedOutput
+            }).ToList() 
+        };
+
+        try
+        {
+            await _appDbContext.Problems.AddAsync(problem);
+            await _appDbContext.SaveChangesAsync();
+            
+        }
+        catch (Exception e)
+        {
+            throw;
+        }
+        
+    }
+
+    public async Task EditProblem(int id, AddProblemDto editProblemDto)
+    {
+        var problem = await _appDbContext.Problems.FirstOrDefaultAsync(p => p.Id == id);
+        
+        if (problem == null)
+        {
+            throw new InvalidOperationException("Problem not found");
+        }
+        
+        var parsedCategory = ParseCategories(editProblemDto.Categories);
+         
+        var parsedDifficulty = ParseDifficulty(editProblemDto.Difficulty);
+        
+        problem.Name = editProblemDto.Name;
+        problem.ProblemText = editProblemDto.ProblemText;
+        problem.Category = parsedCategory;
+        problem.Difficulty = parsedDifficulty;
+        problem.RuntimeLimit = editProblemDto.RuntimeLimit;
+        problem.MemoryLimit = editProblemDto.MemoryLimit;
+        problem.TestCases = editProblemDto.TestCases.Select(tc => new TestCase
+        {
+            Input = tc.Input,
+            ExpectedOutput = tc.ExpectedOutput
+        }).ToList();
+
+
+        try
+        {
+            await _appDbContext.SaveChangesAsync();
+            
+        }
+        catch (Exception ex)
+        {
+            throw new Exception(ex.Message);
+        }
+        
+    }
+    
+    public async Task DeleteProblem(int id)
+    { 
+        var problem = await _appDbContext.Problems.FirstOrDefaultAsync(p => p.Id == id);
+        
+        if (problem == null)
+        {
+            throw new InvalidOperationException("Problem not found");
+        }
+
+        try
+        {
+            _appDbContext.Problems.Remove(problem);
+            await _appDbContext.SaveChangesAsync();
+            
+        }
+        catch (Exception ex)
+        {
+            throw new Exception(ex.Message);
+        }
+    }
 
     public async Task<List<GetSubmissionsResponseDto>> GetSubmissions(int problemId)
     {
@@ -102,5 +199,35 @@ public class ProblemsService
         }).ToList();
         
         return getSubmissions;
+    }
+    
+    
+    private static  Category ParseCategories(List<string> categoryStrings)
+    {
+        Category categories = Category.None;
+
+        foreach (var categoryStr in categoryStrings)
+        {
+            if (Enum.TryParse<Category>(categoryStr, out var category))
+            {
+                categories |= category; // Add to the bitmask
+            }
+            else
+            {
+                throw new ArgumentException($"Invalid category: {categoryStr}");
+            }
+        }
+
+        return categories;
+    }
+    
+    private static Difficulty ParseDifficulty(string difficulty)
+    {
+        if (Enum.TryParse<Difficulty>(difficulty, out var parsedDifficulty))
+        {
+            return parsedDifficulty;
+        }
+
+        return Difficulty.Easy;
     }
 }
